--- src/component/login/Contants.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Contants = () => {
-  return <div></div>;
-};
-
-export default Contants;
