+++ src/pages/Login.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginPage from "../component/login/Screen";
+import LoginPage from "../component/login/LoginScreen";
 
 const Login = () => {
   return <LoginPage />;
